gaps.pl: identify times when text occurs in logs
Count the number of times some text occurs in a log file, and
group those when they are within a certain time of each other. So
if the gap between occurrances is < 5m, for example, count them
together. If > 5m, display them as a separate group. Useful for
identifying when a symptom is occurring and how frequently.