dump-keys-sqlite3: dump active keys as hex
Uses sqlite3 CLI to ensure compatibility. Gets list of active
vbuckets from the master table. Prints this list out as the first
line.  Dumps all keys from all of those vbuckets, encoded as hex,
each on its own line.

The intent is to be able to ship this list of keys somewhere else
where it can be filtered with a regex, etc. Doing it via sqlite
avoids having to run TAP dump, getting all data as well as keys.

Implication is that changes that are only in the disk write queue
will not be seen. Not considered terribly important, since there
is no locking here anyways, so there's always a race condition
there if clients are actively adding or removing items.