dump-keys: fix no-filter case, parallel queries
Fix a bug in the match logic that printed nothing when no
filter was specified.

Require the -a option to dump all keys. Since this is intended
for piping to a delete script, it seems safer to require a
explicit option for this.

Run 4 subshells in parallel for reading the vbucket tables, one
for each bucket_name-[0-3].mb sqlite file.

Other minor cleanups.