Fix major bug in dump-keys-sqlite3.sh; add TTL opt
Major bug is that running in parallel causes output to be jumbled badly.
Just run serially, it doesn't hurt performance appreciably.

Add a -t option for printing TTL, too. Trick is to get -t and -f FILTER
regex to work together; done by changing "^" in REGEX to be ^\d+\|
(skip the TTL part).